Feature/portfolio hero tidyup (#2)
* added helper files

* moved js to helpers and added correct assets
