+++ src/components/Header/PortfolioHero/PortfolioHero.js
@@ -1,22 +1,44 @@
 import React, { Component } from "react";
 
-import dynamics from "dynamics.js";
-import classie from "classie";
-import imagesLoaded from "imagesloaded";
-import AnimOnScroll from "anime";
-import Masonry from "isotope-layout";
-
 import "../../../assets/css/isolayer.css";
 
-import imgOne from "../../../assets/images/thumbnails/isolayer10.jpg";
+// import imgOne from "../../../assets/images/thumbnails/isolayer10.jpg";
+import imgOne from "../../../assets/images/thumbnails/01.jpg";
+import imgTwo from "../../../assets/images/thumbnails/02.jpg";
+import imgThree from "../../../assets/images/thumbnails/03.jpg";
+import imgFour from "../../../assets/images/thumbnails/04.jpg";
+import imgFive from "../../../assets/images/thumbnails/05.jpg";
+import imgSix from "../../../assets/images/thumbnails/06.jpg";
+import imgSeven from "../../../assets/images/thumbnails/07.jpg";
+import imgEight from "../../../assets/images/thumbnails/08.jpg";
+import imgNine from "../../../assets/images/thumbnails/09.jpg";
+import imgTen from "../../../assets/images/thumbnails/10.jpg";
+import imgEleven from "../../../assets/images/thumbnails/11.jpg";
+import imgTwelve from "../../../assets/images/thumbnails/12.jpg";
+
+import {
+  IsoGrid,
+  isoGridProperties,
+  isoGridOptions2,
+  animationHelpers,
+  imagesLoadedHelper
+} from "./PortfolioHeroHelpers";
+
+import {
+  isoGridOptions,
+  isoGridInitEvents,
+  isoGridExpandSubItems,
+  isoGridCollapseSubItems,
+  isoGridScrollPage,
+  isoGridCreatePseudoScroller
+} from "./isoGridHelpers";
 
 class PortfolioHero extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      classes: ["isolayer", "isolayer--deco1", "isolayer--shadow"],
-      updated: false
+      classes: ["isolayer", "isolayer--deco1", "isolayer--shadow"]
     };
   }
 
@@ -24,378 +46,59 @@ class PortfolioHero extends Component {
     // debugger;
     console.log("portfolio hero - componentDidMount()");
 
-    function getComputedTranslateY(obj) {
-      if (!window.getComputedStyle) return;
-      var style = getComputedStyle(obj),
-        transform =
-          style.transform || style.webkitTransform || style.mozTransform;
-      var mat = transform.match(/^matrix3d\((.+)\)$/);
-      if (mat) return parseFloat(mat[1].split(", ")[13]);
-      mat = transform.match(/^matrix\((.+)\)$/);
-      return mat ? parseFloat(mat[1].split(", ")[5]) : 0;
-    }
-
-    var lastTime = 0;
-    var prefixes = "webkit moz ms o".split(" ");
-    // get unprefixed rAF and cAF, if present
-    var requestAnimationFrame = window.requestAnimationFrame;
-    var cancelAnimationFrame = window.cancelAnimationFrame;
-    // loop through vendor prefixes and get prefixed rAF and cAF
-    var prefix;
-    for (var i = 0; i < prefixes.length; i++) {
-      if (requestAnimationFrame && cancelAnimationFrame) {
-        break;
-      }
-      prefix = prefixes[i];
-      requestAnimationFrame =
-        requestAnimationFrame || window[prefix + "RequestAnimationFrame"];
-      cancelAnimationFrame =
-        cancelAnimationFrame ||
-        window[prefix + "CancelAnimationFrame"] ||
-        window[prefix + "CancelRequestAnimationFrame"];
-    }
-
-    if (!requestAnimationFrame || !cancelAnimationFrame) {
-      requestAnimationFrame = function(callback, element) {
-        var currTime = new Date().getTime();
-        var timeToCall = Math.max(0, 16 - (currTime - lastTime));
-        var id = window.setTimeout(function() {
-          callback(currTime + timeToCall);
-        }, timeToCall);
-        lastTime = currTime + timeToCall;
-        return id;
-      };
-
-      cancelAnimationFrame = function(id) {
-        window.clearTimeout(id);
-      };
-    }
-
-    console.log("about to set state");
-
-    var docElem = window.document.documentElement;
-
-    // some helper functions
-    function scrollY() {
-      return window.pageYOffset || docElem.scrollTop;
-    }
-
-    function extend(a, b) {
-      for (var key in b) {
-        if (b.hasOwnProperty(key)) {
-          a[key] = b[key];
-        }
-      }
-      return a;
-    }
+    animationHelpers(window);
 
     const please = () => {
-      console.log("please");
       this.setState(prevState => {
         const currentClasses = prevState.classes;
         currentClasses.push("visible");
-
         return {
-          classes: currentClasses,
-          updated: true
+          classes: currentClasses
         };
       });
     };
 
-    function IsoGrid(el, options) {
-      this.isolayerEl = el;
-
-      this.options = extend({}, this.options);
-      extend(this.options, options);
-
-      if (!document.querySelector(".gridy")) {
-      } else {
-        this.gridEl = this.isolayerEl.querySelector(".gridy");
-      }
-
-      if (!document.querySelector(".grid__item")) {
-      } else {
-        // grid items
-        this.gridItems = [].slice.call(
-          this.gridEl.querySelectorAll(".grid__item")
-        );
-        this.gridItemsTotal = this.gridItems.length;
-
-        this.didscroll = false;
-
-        this._init();
-
-        //please();
-      }
-    }
-
-    IsoGrid.prototype.options = {
-      // static or scrollable
-      type: "static",
-      // grid perspective value
-      perspective: 0,
-      // grid transform
-      transform: "",
-      // each grid item animation (for the subitems)
-      stackItemsAnimation: {
-        // this follows the dynamics.js (https://github.com/michaelvillar/dynamics.js) animate fn syntax
-        // properties (pos is the current subitem position)
-        properties: function(pos) {
-          return {
-            translateZ: (pos + 1) * 50
-          };
-        },
-        // animation options (pos is the current subitem position); itemstotal is the total number of subitems
-        options: function(pos, itemstotal) {
-          return {
-            type: dynamics.bezier,
-            duration: 500,
-            points: [
-              {
-                x: 0,
-                y: 0,
-                cp: [
-                  {
-                    x: 0.2,
-                    y: 1
-                  }
-                ]
-              },
-              {
-                x: 1,
-                y: 1,
-                cp: [
-                  {
-                    x: 0.3,
-                    y: 1
-                  }
-                ]
-              }
-            ]
-          };
-        }
-      },
-      // callback for loaded grid
-      onGridLoaded: function() {
-        return false;
-      }
-    };
+    IsoGrid.prototype.options = isoGridOptions;
 
     IsoGrid.prototype._init = function() {
-      var self = this;
-
-      imagesLoaded(this.gridEl, function() {
-        // initialize masonry
-        self.msnry = new Masonry(self.gridEl, {
-          itemSelector: ".grid__item",
-          isFitWidth: true
-        });
-
-        // the isolayer div element will be positioned fixed and will have a transformation based on the values defined in the HTML (data-attrs for the isolayer div element)
-        if (self.options.type === "scrollable") {
-          self.isolayerEl.style.position = "fixed";
-        }
-
-        self.isolayerEl.style.WebkitTransformStyle = self.isolayerEl.style.transformStyle =
-          "preserve-3d";
-
-        var transformValue =
-          self.options.perspective != 0
-            ? "perspective(" +
-              self.options.perspective +
-              "px) " +
-              self.options.transform
-            : self.options.transform;
-        self.isolayerEl.style.WebkitTransform = self.isolayerEl.style.transform = transformValue;
-
-        // create the div element that will force the height for scrolling
-        if (self.options.type === "scrollable") {
-          self._createPseudoScroller();
-        }
-
-        // init/bind events
-        self._initEvents();
-
-        // effects for loading grid elements:
-        if (self.options.type === "scrollable") {
-          new AnimOnScroll(self.gridEl, {
-            minDuration: 1,
-            maxDuration: 1.2,
-            viewportFactor: 0
-          });
-        }
-
-        // grid is "loaded" (all images are loaded)
-        self.options.onGridLoaded();
-        classie.add(self.gridEl, "grid--loaded");
-      });
+      imagesLoadedHelper(this.gridEl, this);
 
       IsoGrid.prototype._initEvents = function() {
-        var self = this;
-
-        var scrollHandler = function() {
-            requestAnimationFrame(function() {
-              if (!self.didscroll) {
-                self.didscroll = true;
-                self._scrollPage();
-              }
-            });
-          },
-          mouseenterHandler = function(ev) {
-            self._expandSubItems(ev.target);
-          },
-          mouseleaveHandler = function(ev) {
-            self._collapseSubItems(ev.target);
-          };
-
-        if (this.options.type === "scrollable") {
-          // update the transform (ty) on scroll
-          window.addEventListener("scroll", scrollHandler, false);
-          // on resize (layoutComplete for the masonry instance) recalculate height
-          this.msnry.on("layoutComplete", function(laidOutItems) {
-            // reset the height of the pseudoScroller (grid´s height + additional space between the top of the rotated isolayerEl and the page)
-            self.pseudoScrollerEl.style.height =
-              self.gridEl.offsetHeight +
-              self.isolayerEl.offsetTop * Math.sqrt(2) +
-              "px";
-            self._scrollPage();
-          });
-        }
-
-        this.gridItems.forEach(function(item) {
-          item.addEventListener("mouseenter", mouseenterHandler);
-          item.addEventListener("mouseleave", mouseleaveHandler);
-        });
+        isoGridInitEvents(this);
         please();
       };
       IsoGrid.prototype._expandSubItems = function(item) {
-        var self = this,
-          itemLink = item.querySelector("a"),
-          subItems = [].slice.call(itemLink.querySelectorAll(".layer")),
-          subItemsTotal = subItems.length;
-
-        itemLink.style.zIndex = item.style.zIndex = this.gridItemsTotal;
-
-        subItems.forEach(function(subitem, pos) {
-          dynamics.stop(subitem);
-          dynamics.animate(
-            subitem,
-            self.options.stackItemsAnimation.properties(pos),
-            self.options.stackItemsAnimation.options(pos, subItemsTotal)
-          );
-        });
+        isoGridExpandSubItems(this, item);
       };
 
       IsoGrid.prototype._collapseSubItems = function(item) {
-        var itemLink = item.querySelector("a");
-        [].slice
-          .call(itemLink.querySelectorAll(".layer"))
-          .forEach(function(subitem, pos) {
-            dynamics.stop(subitem);
-            dynamics.animate(
-              subitem,
-              {
-                translateZ: 0 // enough to reset any transform value previously set
-              },
-              {
-                duration: 100,
-                complete: function() {
-                  itemLink.style.zIndex = item.style.zIndex = 1;
-                }
-              }
-            );
-          });
+        isoGridCollapseSubItems(item);
       };
 
       IsoGrid.prototype._scrollPage = function() {
-        this.gridEl.style.WebkitTransform = this.gridEl.style.transform =
-          "translate3d(0,-" + scrollY() + "px,0)";
-        this.didscroll = false;
+        isoGridScrollPage(this);
       };
     };
 
     IsoGrid.prototype._createPseudoScroller = function() {
-      // element that will force the height for scrolling
-      this.pseudoScrollerEl = document.createElement("div");
-      this.pseudoScrollerEl.className = "pseudo-scroller";
-      // insert it inside the main container (same level of isolayerEl)
-      this.isolayerEl.parentNode.insertBefore(
-        this.pseudoScrollerEl,
-        this.isolayerEl
-      );
-      // set the height of the pseudoScroller (grid´s height + additional space between the top of the rotated isolayerEl and the page - value set for the translation on the Y axis)
-      this.pseudoScrollerEl.style.height =
-        this.gridEl.offsetHeight +
-        getComputedTranslateY(this.isolayerEl) * Math.sqrt(2) +
-        "px";
+      isoGridCreatePseudoScroller(this);
     };
-
-    function getRandomInt(min, max) {
-      return Math.floor(Math.random() * (max - min + 1)) + min;
-    }
 
     new IsoGrid(document.querySelector(".isolayer--deco1"), {
       transform:
         "translateX(33vw) translateY(-340px) rotateX(45deg) rotateZ(45deg)",
       stackItemsAnimation: {
-        properties: function(pos) {
-          return {
-            translateZ: (pos + 1) * 30,
-            rotateZ: getRandomInt(-4, 4)
-          };
-        },
-        options: function(pos, itemstotal) {
-          return {
-            type: dynamics.bezier,
-            duration: 500,
-            points: [
-              {
-                x: 0,
-                y: 0,
-                cp: [
-                  {
-                    x: 0.2,
-                    y: 1
-                  }
-                ]
-              },
-              {
-                x: 1,
-                y: 1,
-                cp: [
-                  {
-                    x: 0.3,
-                    y: 1
-                  }
-                ]
-              }
-            ],
-            delay: (itemstotal - pos - 1) * 40
-          };
-        }
+        properties: isoGridProperties,
+        options: isoGridOptions2
       }
     });
-    // this.setState(prevState => {
-    //   const currentClasses = prevState.classes;
-    //   currentClasses.push("visible");
-
-    //   return {
-    //     classes: currentClasses,
-    //     updated: true
-    //   };
-    // });
   }
 
-  // shouldComponentUpdate() {
-  //   return false;
-  // }
   render() {
     console.log("portfolio hero - render()");
 
     return (
       <div>
-        {this.state.updated ? <p>loaded</p> : <p>loading</p>}
-
         <div className="inner">
           <h2>WILL SHAW</h2>
           <p>Videographer | Editor | Photographer</p>
@@ -435,7 +138,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgTwo} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -443,7 +146,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgThree} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -451,7 +154,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgFour} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -459,7 +162,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgFive} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -467,7 +170,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgSix} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -475,7 +178,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img className="grid__img layer" src={imgOne} alt="02" />
+                <img className="grid__img layer" src={imgSeven} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -483,11 +186,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="../../../assets/images/thumbnails/Changin-Plants-4-300x188.png"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgEight} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -495,11 +194,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer9.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgNine} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -507,11 +202,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer10.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgTen} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -519,11 +210,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer11.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgEleven} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -531,11 +218,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer12.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgTwelve} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -543,11 +226,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer13.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgOne} alt="image" />
               </a>
             </li>
             <li className="grid__item">
@@ -555,11 +234,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer14.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgTwo} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -567,11 +242,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer15.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgThree} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -579,11 +250,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer16.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgFour} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -591,11 +258,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer17.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgFive} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -603,11 +266,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer18.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgThree} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -615,11 +274,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer19.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgFour} alt="02" />
               </a>
             </li>
             <li className="grid__item">
@@ -627,11 +282,7 @@ class PortfolioHero extends Component {
                 <div className="layer" />
                 <div className="layer" />
                 <div className="layer" />
-                <img
-                  className="grid__img layer"
-                  src="images/isolayer20.jpg"
-                  alt="02"
-                />
+                <img className="grid__img layer" src={imgFive} alt="02" />
               </a>
             </li>
           </ul>
